requirements.txt
markdown-poster@ git+https://github.com/Lightning-AI/LAI-markdown-poster-Component.git@760ddecb73f65af19a6968cbc81df651dac586fe
mkposters@ git+https://github.com/stanbiryukov/mkposters@522858cfe06d5438094302ff3d5c8b9256a5a4de
protobuf==3.20.1
rich==12.4.1
torch==1.11.0
torchvision==0.12.0
gradio==2.9.4
transformers==4.19.2
streamlit==1.9.0
streamlit-autorefresh==0.0.1
nbconvert==6.5.0
jupyterlite==0.1.0b7
lightning
jupyterlab==3.3.3
imageio-ffmpeg==0.4.7
